Break types into their own file.
