+++ src/Workspaces/Core/Portable/Packaging/IPackageSearchService.cs
@@ -1,11 +1,7 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Host;
 
 namespace Microsoft.CodeAnalysis.Packaging
@@ -23,38 +19,6 @@ namespace Microsoft.CodeAnalysis.Packaging
         /// </summary>
         IEnumerable<PackageWithTypeResult> FindPackagesWithType(
             string source, string name, int arity, CancellationToken cancellationToken);
-    }
-
-    internal interface IReferenceAssemblySearchService : IWorkspaceService
-    {
-        /// <summary>
-        /// Searches for reference assemblies that contain a type with the provided name and arity.
-        /// Note: Implementations are free to return the results they feel best for the
-        /// given data.  Specifically, they can do exact or fuzzy matching on the name.
-        /// They can use or ignore the arity depending on their capabilities. 
-        /// 
-        /// Implementations should return results in order from best to worst (from their
-        /// perspective).
-        /// </summary>
-        IEnumerable<ReferenceAssemblyWithTypeResult> FindReferenceAssembliesWithType(
-            string name, int arity, CancellationToken cancellationToken);
-    }
-
-    internal class ReferenceAssemblyWithTypeResult
-    {
-        public readonly IReadOnlyList<string> ContainingNamespaceNames;
-        public readonly string AssemblyName;
-        public readonly string TypeName;
-
-        public ReferenceAssemblyWithTypeResult(
-            string assemblyName,
-            string typeName,
-            IReadOnlyList<string> containingNamespaceNames)
-        {
-            AssemblyName = assemblyName;
-            TypeName = typeName;
-            ContainingNamespaceNames = containingNamespaceNames;
-        }
     }
 
     internal class PackageWithTypeResult
